requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
openai==1.30.1
python-multipart==0.0.9
PyYAML==6.0.1
anyio==4.3.0
httpx==0.27.0